requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.11
discord.py==1.7.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
gspread==5.1.1
idna==3.3
multidict==6.0.2
numpy==1.22.2
oauthlib==3.2.0
pandas==1.4.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
typing_extensions==4.0.1
urllib3==1.26.8
yarl==1.7.2
